requirements.txt
Django
Pillow
qrcode[pil]
requests
django-environ
